requirements.txt
h5py==2.9.0
Flask==1.1.2
numpy==1.18.5
opencv-python==4.4.0.46
Pillow==7.1.2
scikit-learn==0.23.2
tensorflow==2.1.2
tqdm==4.53.0
